pyproject.toml
[tool.poetry]
name = "homework-13.2-web"
version = "0.1.0"
description = ""
authors = ["fls0 <134416720+fls0@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
psycopg2 = "^2.9.9"
django-taggit = "^5.0.1"
pymongo = "^4.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
